gavicore/pyproject.toml
#  Copyright (c) 2025 by the Eozilla team and contributors
#  Permissions are hereby granted under the terms of the Apache 2.0 License:
#  https://opensource.org/license/apache-2-0.

[project]
name = "gavicore"
version = "0.1.0.dev0"
description = "Pydantic data models and common utilities for other Eozilla packages"
readme = "README.md"
authors = [
  {name = "Norman Fomferra (Brockmann Consult GmbH)"},
  {name = "Tejas Morbagal Harish (Brockmann Consult GmbH)"}
]
keywords = [
  "eo",
  "esa",
  "data processing",
  "ogc",
]
license = {text = "Apache-2.0"}
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX",
  "Operating System :: Unix",
  "Operating System :: MacOS",
]
requires-python = ">=3.10"
dependencies = [
  "click",
  # "panel",  # TODO: declare as optional dependency
  "pydantic",
  "pyyaml",
  "typer",
]

[project.urls]
Documentation = "https://eo-tools.github.io/eozilla"
Repository = "https://github.com/eo-tools/eozilla"
Changelog = "https://github.com/eo-tools/eozilla/blob/main/CHANGES.md"
Issues = "https://github.com/eo-tools/eozilla/issues"

[build-system]
requires = ["hatch"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.sdist]
include = [
  "src/gavicore/**/*",
  "pyproject.toml",
  "README.md",
  "LICENSE",
]

[tool.hatch.build.targets.wheel]
# packages = ["gavicore"]
# include = ["src/gavicore/**/*"]
